Implemented basic logic in repository and model layers

+++ lavugio-mobile/app/src/main/java/com/example/lavugio_mobile/data/model/reports/ChartData.java
@@ -7,14 +7,14 @@ public class ChartData {
     private String xAxisLabel;
     private String yAxisLabel;
     private List<String> labels;
-    private List<Float> data;
-    private float sum;
-    private float average;
+    private List<Double> data;
+    private double sum;
+    private double average;
 
     public ChartData() {}
 
     public ChartData(String title, String xAxisLabel, String yAxisLabel,
-                     List<String> labels, List<Float> data, float sum, float average) {
+                     List<String> labels, List<Double> data, double sum, double average) {
         this.title = title;
         this.xAxisLabel = xAxisLabel;
         this.yAxisLabel = yAxisLabel;
@@ -32,10 +32,10 @@ public class ChartData {
     public void setYAxisLabel(String yAxisLabel) { this.yAxisLabel = yAxisLabel; }
     public List<String> getLabels() { return labels; }
     public void setLabels(List<String> labels) { this.labels = labels; }
-    public List<Float> getData() { return data; }
-    public void setData(List<Float> data) { this.data = data; }
-    public float getSum() { return sum; }
-    public void setSum(float sum) { this.sum = sum; }
-    public float getAverage() { return average; }
-    public void setAverage(float average) { this.average = average; }
+    public List<Double> getData() { return data; }
+    public void setData(List<Double> data) { this.data = data; }
+    public double getSum() { return sum; }
+    public void setSum(double sum) { this.sum = sum; }
+    public double getAverage() { return average; }
+    public void setAverage(double average) { this.average = average; }
 }
